Render handles in View and Edit Policies
Summary:
Fixes T3184
We make sure that we're not working with a global policy that just sets the title correctly. Otherwise we have a PHID to work with and set the handle as required.

Test Plan: Change policies on a mock and see the title render correctly.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, Korvin

Maniphest Tasks: T3184

Differential Revision: https://secure.phabricator.com/D6042