Let requestors claim their code is picked
Summary:
In Releeph, `ReleephRequestHeaderView` will render you a "mark-manually-picked" button if (a) you are authoritative (chuckr) or (b) you are the person who created this Releeph request.

However `ReleephRequestActionController`, which handles these button presses, will only do something if you are (a).  This patch honors your button pressing if you are (b) as well.

Test Plan: Push buttons.  There's another bug in the Javascript do-stuff-without-reloading-the-page code that handles these button pushes, but I'd like to fix that separately from handling this hi-pri bug.

Reviewers: epriestley, wez

Reviewed By: epriestley

CC: aran, Korvin

Differential Revision: https://secure.phabricator.com/D6036