Conpherence notifications - fix ordering
Summary: we get participation data ordered, then query conpherences by phid... be sure to resort the conpherences based on participation data. I missed this in testing 'cuz my test data is so trashy, but it is glaringly obvious in production. :/

Test Plan: replied to a very old conpherence and noted it was first in the notification panel

Reviewers: chad, epriestley

Reviewed By: epriestley

CC: Korvin, aran

Maniphest Tasks: T3641

Differential Revision: https://secure.phabricator.com/D6711