Add `src/extensions/` to Phabricator
Summary: I added this easier-extension mechanism a while ago but only added the actual directories to libphutil and arcanist. This one should work, it just wasn't checked into the repository.

Test Plan: Added a file with an exception in it to the directory, verified the exception was thrown at runtime.

Reviewers: btrahan

Reviewed By: btrahan

CC: aran

Differential Revision: https://secure.phabricator.com/D6759